+++ slang/lib/builder/builder/slang_file_collection_builder.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 
 import 'package:collection/collection.dart';
 import 'package:slang/builder/builder/raw_config_builder.dart';
+import 'package:slang/builder/decoder/base_decoder.dart';
 import 'package:slang/builder/model/i18n_locale.dart';
 import 'package:slang/builder/model/raw_config.dart';
 import 'package:slang/builder/model/slang_file_collection.dart';